Use Long's compared to Integers

+++ src/main/java/io/github/haappi/library/Utils.java
@@ -5,12 +5,12 @@ public class Utils {
         throw new UnsupportedOperationException("This is a utility class and cannot be instantiated");
     }
 
-    public static Integer getRandomStringOfNumbers(Integer length) {
+    public static Long getRandomStringOfNumbers(Integer length) {
         StringBuilder sb = new StringBuilder();
         for (int i = 0; i < length; i++) {
             sb.append(getRandomNumber(1, 10));
         }
-        return Integer.parseInt(sb.toString());
+        return Long.parseLong(sb.toString());
     }
 
     public static Integer getRandomNumber(Integer min, Integer max) {
